move events to common folder
