+++ components/shopify/shopify.app.mjs
@@ -2,171 +2,8 @@ import get from "lodash.get";
 import Shopify from "shopify-api-node";
 import toPath from "lodash.topath";
 import retry from "async-retry";
+import { events } from "./sources/common/events.mjs";
 import { toSingleLineString } from "./actions/commons.mjs";
-
-const events = [
-  {
-    label: "Article Created",
-    value: JSON.stringify({
-      filter: "Article",
-      verb: "create",
-    }),
-  },
-  {
-    label: "Article Destroyed",
-    value: JSON.stringify({
-      filter: "Article",
-      verb: "destroy",
-    }),
-  },
-  {
-    label: "Article Published",
-    value: JSON.stringify({
-      filter: "Article",
-      verb: "published",
-    }),
-  },
-  {
-    label: "Article Unpublished",
-    value: JSON.stringify({
-      filter: "Article",
-      verb: "unpublished",
-    }),
-  },
-  {
-    label: "Article Updated",
-    value: JSON.stringify({
-      filter: "Article",
-      verb: "update",
-    }),
-  },
-  {
-    label: "Blog Created",
-    value: JSON.stringify({
-      filter: "Blog",
-      verb: "create",
-    }),
-  },
-  {
-    label: "Blog Destroyed",
-    value: JSON.stringify({
-      filter: "Blog",
-      verb: "destroy",
-    }),
-  },
-  {
-    label: "Collection Created",
-    value: JSON.stringify({
-      filter: "Collection",
-      verb: "create",
-    }),
-  },
-  {
-    label: "Collection Destroyed",
-    value: JSON.stringify({
-      filter: "Collection",
-      verb: "destroy",
-    }),
-  },
-  {
-    label: "Collection Published",
-    value: JSON.stringify({
-      filter: "Collection",
-      verb: "published",
-    }),
-  },
-  {
-    label: "Collection Unpublished",
-    value: JSON.stringify({
-      filter: "Collection",
-      verb: "unpublished",
-    }),
-  },
-  {
-    label: "Order Confirmed",
-    value: JSON.stringify({
-      filter: "Order",
-      verb: "confirmed",
-    }),
-  },
-  {
-    label: "Page Created",
-    value: JSON.stringify({
-      filter: "Page",
-      verb: "create",
-    }),
-  },
-  {
-    label: "Page Destroyed",
-    value: JSON.stringify({
-      filter: "Page",
-      verb: "destroy",
-    }),
-  },
-  {
-    label: "Page Published",
-    value: JSON.stringify({
-      filter: "Page",
-      verb: "published",
-    }),
-  },
-  {
-    label: "Page Unpublished",
-    value: JSON.stringify({
-      filter: "Page",
-      verb: "unpublished",
-    }),
-  },
-  {
-    label: "Price Rule Created",
-    value: JSON.stringify({
-      filter: "PriceRule",
-      verb: "create",
-    }),
-  },
-  {
-    label: "Price Rule Destroyed",
-    value: JSON.stringify({
-      filter: "PriceRule",
-      verb: "destroy",
-    }),
-  },
-  {
-    label: "Price Rule Updated",
-    value: JSON.stringify({
-      filter: "PriceRule",
-      verb: "update",
-    }),
-  },
-  {
-    label: "Product Created",
-    value: JSON.stringify({
-      filter: "Product",
-      verb: "create",
-    }),
-  },
-  {
-    label: "Product Destroyed",
-    value: JSON.stringify({
-      filter: "Product",
-      verb: "destroy",
-    }),
-  },
-  {
-    label: "Product Published",
-    value: JSON.stringify({
-      filter: "Product",
-      verb: "published",
-    }),
-  },
-  {
-    label: "Product Unpublished",
-    value: JSON.stringify({
-      filter: "Product",
-      verb: "unpublished",
-    }),
-  },
-];
 
 export default {
   type: "app",
